requirements.txt
flask
mysql-connector-python
flask-jwt-extended
python-dotenv
bcrypt
pyjwt
